Proof-of-concept: Move the nmethod header out of code cache
The structure of an nmethod object in the code cache before this patch
looks something like this -

|CodeBlob header| |NMethod header| |Constants| |MainCode| |StubCode|
|Data/oops|

Mutable and immutable data have already been moved out of code cache and
only pointers to that data are stored in the NMethod header.

The total header space occupied in an nmethod object (JDK25) is 224B.

This patch helps remove the NMethod header out of codecache which is
around 160B (while the CodeBlob header is 64B) and only adds a pointer
to the header structure in the C heap bringing the total NMethod header
to 72B (64B CodeBlob header + 8B pointer to NMethod header) which is a
~3x reduction in nmethod header size.

The structure of an nmethod object now looks like this after this patch:

|CodeBlob header| |Ptr to NMethodHeader| |Constants| |MainCode| |StubCode|
|Data/oops|

Note: After this change, the templateTable_<cpuType>.cpp need to be
modified to add an extra level of indirection while accessing _state and
_osr_entry_point fields of the NMethodHeader. I have only modified these
files for aarch64 and x86 for now. This patch might fail to build on any
other platform.

Change-Id: Ib43e5e8f5a52afb09a1e188e2adb624806689d8c